+++ frontend/src/pages/game/[gameId].module.scss
@@ -1,6 +1,6 @@
 .layout {
 	display: flex;
 	padding: 1rem;
-	gap: 1rem;
+	gap: 2rem;
 	margin: 1rem 2rem;
 }
